refactor(api): extract demo-tracking data layer + address review panel
- Extract SQL + fold/join logic to lib/demo-tracking.ts (below the route
  layer, directly unit/-pg testable). Route is now a thin Hono/Effect shell.
- New platform-demo-pg.test.ts: real-Postgres test locking in the text=uuid
  join (token_usage.conversation_id::text = conversations.id) + uuid[] bind.
- New demo-tracking.test.ts: pure unit tests for the orphaned-usage-row drop,
  mixed priced/unpriced partial cost, and count-weighted latency branches.
- foldUsage: Number.isFinite-guarded bigint-string coercion (NaN -> 0).
- Full demo.ts mock export surface in the route test (mock-all-exports).
- agent-token-usage: assert latency is an integer (units/formula regression).
- extractText: guard JSON.stringify so unrenderable content can't crash the
  Sheet; tighten the doc. Web DemoConfigSchema -> .int() to match the API.
- token-pricing: fix 'inputTokens.total' -> 'inputTokens' comment (no .total).

+++ packages/web/src/ui/lib/admin-schemas.ts
@@ -543,8 +543,8 @@ export type {
 
 export const DemoConfigSchema = z.object({
   model: z.string(),
-  maxSteps: z.number(),
-  rpm: z.number(),
+  maxSteps: z.number().int(),
+  rpm: z.number().int(),
   effectiveModel: z.string().nullable(),
 });
 export type DemoConfig = z.infer<typeof DemoConfigSchema>;
